# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Monitoring platform 1000 Moustaches API",
    "description": "API documentation for the 1000 Moustaches monitoring platform",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/animal-host-families/hostFamily/{hostFamilyId}": {
      "get": {
        "tags": [
          "Host Family Relations"
        ],
        "summary": "Get animals hosted by host family",
        "description": "Retrieve all animals that have been hosted by a specific host family",
        "parameters": [
          {
            "name": "hostFamilyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Host family ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of animals hosted by the host family",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimalHostFamily"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/species/": {
      "get": {
        "tags": [
          "Species"
        ],
        "summary": "Get all species",
        "description": "Retrieve a list of all animal species",
        "responses": {
          "200": {
            "description": "List of species",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Species"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Species"
        ],
        "summary": "Create new species",
        "description": "Create a new animal species",
        "responses": {
          "201": {
            "description": "Species created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Species"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Species"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Cat"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/species/{id}": {
      "get": {
        "tags": [
          "Species"
        ],
        "summary": "Get species by ID",
        "description": "Retrieve a specific species by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Species ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Species details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Species"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Species"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Species"
        ],
        "summary": "Update species",
        "description": "Update an existing species",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Species ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Species updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Species"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Species"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Updated Cat"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Species"
        ],
        "summary": "Delete species",
        "description": "Delete a species by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Species ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Species deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/hostFamily/{hostFamilyId}": {
      "get": {
        "tags": [
          "Host Family Relations"
        ],
        "summary": "Get animals hosted by host family",
        "description": "Retrieve all animals that have been hosted by a specific host family",
        "parameters": [
          {
            "name": "hostFamilyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Host family ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of animals hosted by the host family",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimalHostFamily"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Authorization": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Firebase authentication token"
      }
    },
    "schemas": {
      "Species": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Species name",
            "example": "Chat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
...
      "Animal": {
        "type": "object",
        "required": [
          "name",
          "icad"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Animal name",
            "example": "Whiskers"
          },
          "species": {
            "$ref": "#/components/schemas/Species"
          },
          "icad": {
            "type": "string",
            "description": "ICAD identification number",
            "example": "250268500123456"
          },
          "sexe": {
            "type": "string",
            "description": "Animal gender",
            "enum": [
              "M",
              "F"
            ],
            "example": "F"
          },
          "race": {
            "type": "string",
            "description": "Animal breed",
            "example": "European Shorthair"
          },
          "adopted": {
            "type": "boolean",
            "description": "Is adopted",
            "default": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
      "HostFamily": {
        "type": "object",
        "required": [
          "name",
          "firstname"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Host family last name",
            "example": "Martin"
          },
          "firstname": {
            "type": "string",
            "description": "Host family first name",
            "example": "Marie"
          },
          "phone": {
            "type": "string",
            "description": "Phone number",
            "example": "+33123456789"
          },
          "mail": {
            "type": "string",
            "format": "email",
            "description": "Email address",
            "example": "marie.martin@example.com"
          },
          "address": {
            "type": "string",
            "description": "Address",
            "example": "456 Avenue des Fleurs, 69000 Lyon"
          },
          "isAvailable": {
            "type": "boolean",
            "description": "Is available for hosting",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
...
      "AnimalHostFamily": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier",
            "example": 1
          },
          "animal": {
            "$ref": "#/components/schemas/Animal"
          },
          "hostFamily": {
            "$ref": "#/components/schemas/HostFamily"
          },
          "entryDate": {
            "type": "string",
            "format": "date-time",
            "description": "Date when animal entered the host family",
            "example": "2023-12-15T10:30:00Z"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "Authorization": []
    }
  ]
}